pyproject.toml
[tool.poetry]
name = "openai-discord-bot"
version = "0.1.0"
description = ""
authors = ["loc <l0c4t0r@gmail.com>"]
readme = "README.md"
packages = [{include = "openai_discord_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.2.2"
python-dotenv = "^1.0.0"
httpx = "^0.23.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
